Project cleanup!
- Moved functionality pieces into their own components. Modularity makes maintenance easier.
