+++ FirstProject/taskopedia/src/components/Footer.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Footer() {
+  return <footer style={{ color: "gray", backgroundColor: "black" }}>Happy Coding!</footer>;
+}
+
+export default Footer;
